% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{concentration_measurements}
\alias{concentration_measurements}
\title{Model concentration measurements}
\usage{
concentration_measurements(
  measurements = NULL,
  id_col = "sample_id",
  concentration_col = "concentration",
  replicate_col = NULL,
  n_averaged = 1,
  n_averaged_col = NULL,
  total_partitions_col = NULL,
  distribution = "gamma"
)
}
\arguments{
\item{measurements}{A \code{data.frame} with each row representing one dPCR
measurement. Must have at least a column with sample ids and a column with
concentration measurements.}

\item{id_col}{Name of the column containing the id of the sample.}

\item{concentration_col}{Name of the column containing the measured
concentrations.}

\item{replicate_col}{Name of the column containing the replicate ID of each
measurement. This is used to identify multiple measurements (biological
replicates) of the same sample. Should be \code{NULL} if only one biological
replicate per sample was made.}

\item{n_averaged}{The number of technical replicates (i.e. repeated PCR runs)
used for each biological replicate. The concentration provided in the
\code{measurements} \code{data.frame} is assumed to be the average or pooled estimate
from several technical replicates. Can be either a single number (it is
then assumed that the number of averaged replicates is the same for each
observation) or a vector (one value for each observation).}

\item{n_averaged_col}{Name of the column in the \code{measurements} data.frame
containing the number of technical replicates over which the measurements
have been averaged/pooled. This is an alternative to specifying
\code{n_averaged}.}

\item{total_partitions_col}{Name of the column in the \code{measurements}
data.frame containing the number of total partitions (e.g. droplets for
ddPCR) in the dPCR reaction of each measurement. If several technical
replicates are used, this should be the average number of valid partitions
per replicate. Only applies when modeling concentration measurements via
the dPCR-specific noise model. Can be used by the \code{\link[=noise_dPCR]{noise_dPCR()}}
and \code{\link[=nondetect_dPCR]{nondetect_dPCR()}} modeling components. Note that this is really the
number of \emph{valid} partitions, not the number of positive partitions.}

\item{distribution}{Parametric distribution for concentration measurements.
Currently supported are "gamma" (default and recommended), "log-normal",
"truncated normal", and "normal". The "truncated normal" and "normal"
options are not recommended for use in practice.}
}
\description{
This option models concentration measurements from digital PCR
and can be used in settings where the numbers of total and positive
partitions are unknown, i.e. only the measured concentrations are provided.
}
