% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCRfit.R
\name{dPCRfit}
\alias{dPCRfit}
\title{Fit a dPCR-specific regression model}
\usage{
dPCRfit(
  formula,
  data,
  link = c("identity", "log"),
  measurements = concentration_measurements(),
  noise = noise_dPCR(),
  nondetect = nondetect_dPCR(),
  prior_intercept = c(0, 1),
  prior_coefficients = c(0, 1),
  fit_opts = set_fit_opts()
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted, similar to a
formula provided to \code{lm}. This is where you specify the covariates.}

\item{data}{A data.frame containing the variables in the model.}

\item{link}{The link function to use. Can be "identity" or "log".}

\item{measurements}{Assumptions about the measurement process and laboratory
parameters, provided through the \code{\link[=concentration_measurements]{concentration_measurements()}} component.}

\item{noise}{Assumptions about the noise in the data, provided through the
\code{\link[=noise_dPCR]{noise_dPCR()}} or the \code{\link[=noise_constant_cv]{noise_constant_cv()}} / \code{\link[=noise_constant_var]{noise_constant_var()}}
component.}

\item{nondetect}{Assumptions about the non-detects in the data, provided
through the \code{\link[=nondetect_dPCR]{nondetect_dPCR()}} or the \code{\link[=nondetect_none]{nondetect_none()}} component.}

\item{prior_intercept}{A vector of two values (mean, sd) specifying the
normal prior for the intercept.}

\item{prior_coefficients}{A vector of two values (mean, sd) specifying the
normal prior for the regression coefficients.}

\item{fit_opts}{A list of options for the fitting process. See
\code{set_fit_opts()} for details.}
}
\value{
An object of class "dPCRfit_result" containing the fitted model and
additional details.
}
\description{
Fits regression models to gene concentration measurements from
digital PCR (dPCR) using a dPCR-specific likelihood that accounts for
concentration-dependent error and non-detects. Supports multiple linear
regression with identity or log link, incorporating prior information on
lab parameters. Models are fitted via Stan using the 'cmdstanr' interface.
}
