% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{dPCRfit_compile}
\alias{dPCRfit_compile}
\title{Compile the dPCRfit model}
\usage{
dPCRfit_compile(force_recompile = FALSE, verbose = FALSE)
}
\arguments{
\item{force_recompile}{If TRUE, the model will be recompiled even if it
has already been successfully compiled.}

\item{verbose}{If TRUE, warnings and detailed errors from the compilation are
printed. This can help to diagnose compilation issues.}
}
\description{
The stan model used by dPCRfit needs to be compiled for your
device. This is only necessary once after installing or updating the
package.
}
\details{
If the model is not successfully compiled, please
ensure that \code{cmdstan} is properly set up and try updating it to a newer
version using \code{\link[cmdstanr:install_cmdstan]{cmdstanr::install_cmdstan()}}. If the problem persists,
please run \code{dPCRfit_compile(verbose = TRUE)} and post the output in
a new issue on GitHub, along with your \code{\link[cmdstanr:set_cmdstan_path]{cmdstanr::cmdstan_version()}}.
}
