% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{model_component}
\alias{model_component}
\title{Create a model_component object}
\usage{
model_component(name, .e, args = as.list(parent.frame()))
}
\arguments{
\item{name}{A character string representing the name of the model component.}

\item{.e}{An expression that defines the model component through changes to
the modeldata object.}

\item{args}{Arguments to be passed to the model component. By default, this
is taken from the calling function.}
}
\value{
An S3 object of class "model_component".
}
\description{
Create a model_component object
}
\keyword{internal}
