% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{model_stan_opts}
\alias{model_stan_opts}
\title{Specify details of the stan model}
\usage{
model_stan_opts(threads = FALSE, force_recompile = FALSE)
}
\arguments{
\item{threads}{Should multihreading be enabled? Default is `FALSE`, as
`dPCRfit` currently does not support within-chain parallelism.}

\item{force_recompile}{Should recompilation be forced before model fitting?
If `FALSE` (default), the model is only recompiled when changes to the
model code are detected. However, as the change detection is not fully
reliable, it is sometimes necessary to force recompilation after having
made changes to the stan code.}
}
\value{
A list with details of the stan model
}
\description{
Specify details of the stan model
}
