% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{modeldata_init}
\alias{modeldata_init}
\title{Construct an unspecified model}
\usage{
modeldata_init()
}
\value{
A `modeldata` object containing data and specifications of the model
  to be fitted. Can be combined with other model components to
  add data and model specifications.

The `modeldata` object also includes information about parameter
  initialization (`.init`), meta data (`.metainfo`), and checks to be
  performed before model fitting (`.checks`).
}
\description{
Construct an unspecified model
}
