% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{noise_constant_cv}
\alias{noise_constant_cv}
\title{Model measurement noise with constant coefficient of variation}
\usage{
noise_constant_cv(cv_prior_mu = 0, cv_prior_sigma = 1)
}
\arguments{
\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements.}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}
}
\description{
This option estimates measurement noise via a constant
coefficient of variation model.

For a non-constant coefficient of variation model, see
\code{\link[=noise_dPCR]{noise_dPCR()}}.
}
\details{
The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements (\code{cv}): \verb{Truncated normal}
}
}
