% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{noise_constant_var}
\alias{noise_constant_var}
\title{Model measurement noise with constant variance}
\usage{
noise_constant_var(cv_prior_mu = 0, cv_prior_sigma = 1, warn = TRUE)
}
\arguments{
\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements.}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}
}
\description{
This option estimates measurement noise via a constant variance
  model.This is a misspecification for dPCR data and is only supported for
  comparison purposes.

For a constant coefficient of variation model, see
  [noise_constant_cv()], and for a non-constant coefficient of
  variation model, see [noise_dPCR()].
}
\details{
Note that although this model keeps the variance constant, the prior
  for the measurement noise is still in terms of the (average) coefficient of
  variation (CV). This makes prior specification easier since the CV is
  unitless.

The priors of this component have the following functional form:
- coefficient of variation of concentration measurements: `Truncated normal`
}
