% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{noise_dPCR}
\alias{noise_dPCR}
\title{Model measurement noise for digital PCR data}
\usage{
noise_dPCR(
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  partitions_observe = FALSE,
  max_partitions = 30000,
  partition_loss_mean = 0.1,
  volume_scaled = 1e-05,
  prePCR_noise_type = "log-normal"
)
}
\arguments{
\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of analyzed
concentrations (pre-PCR noise). Note that this does \emph{not} include the
technical noise of the digital PCR. This is because the dPCR noise is
explicitly modeled (using the total number of partitions and conversion
factor).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of analyzed concentrations.}

\item{partitions_observe}{If TRUE, the total number of partitions is taken
from the supplied measurements \code{data.frame}. This requires that the
argument \code{total_partitions_col} is specified in
\code{\link[=concentration_measurements]{concentration_measurements()}}.}

\item{max_partitions}{The maximum total number of dPCR partitions. This is
usually defined by the manufacturer of the dPCR system/chip used, which
supports a certain maximum number of partitions.}

\item{partition_loss_mean}{The mean relative partition loss. A certain
proportion of partitions in a dPCR run is typically invalid and discarded
from the concentration estimate.}

\item{volume_scaled}{The conversion factor (partition volume multiplied with
the scaling of concentration in the assay) for the dPCR reaction. See
details for further explanation.}

\item{prePCR_noise_type}{The parametric distribution to assume for noise
before the PCR assay. Currently supported are "log-normal" and "gamma". The
choice of the parametric distribution typically makes no relevant
difference for the noise model, but can make a relevant difference for the
LOD model if \code{\link[=nondetect_dPCR]{nondetect_dPCR()}} is used.}
}
\description{
This option models concentration measurements using an error
model specialized for digital PCR (e.g. digital droplet PCR). This is a
fast approximation assuming a fixed number of valid total partitions in the
dPCR assay and a known conversion factor. For a model that jointly
estimates uncertain assay parameters, see \code{\link[=noise_dPCR_params]{noise_dPCR_params()}}.
}
\details{
The conversion factor (see \code{volume_scaled}) is the partition volume
v multiplied with a scaling factor s. The scaling factor accounts for
concentration differences between the sample and the reaction mix, for
example due to extraction or adding of reagents. For example, if the
partition volume is 4.5e-7 mL and the scaling factor is 100:3 (i.e. 100
gc/mL in the original sample correspond to 3 gc/mL in the PCR reaction),
then the overall conversion factor is 4.5e-7 * 100 / 3 = 1.5e-5.

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements (\code{cv}): \verb{Truncated normal}
}
}
