% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{nondetect_none}
\alias{nondetect_none}
\title{Do not model non-detects}
\usage{
nondetect_none()
}
\description{
This option drops all zero measurements from the likelihood.
}
\details{
Dropping zero measurements implicitly conditions on positive
  measurements. This can lead to an upward bias when estimating low
  concentrations.
}
