% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot prior distributions for model components}
\usage{
plot_prior(component, type, n_draws = 1000, show_draws = 50, seed = 0)
}
\arguments{
\item{component}{A model component object. Currently the following components
are supported:
\itemize{
\item \code{\link[=noise_dPCR]{noise_dPCR()}}: dPCR noise model component
}}

\item{type}{Type of prior to plot. Currently the following priors are
supported:
\itemize{
\item component \code{\link[=noise_dPCR]{noise_dPCR()}}: "partitions"
}}

\item{n_draws}{Number of draws to simulate from the prior distributions.}

\item{show_draws}{Number of example draws to show in the plots.}

\item{seed}{Seed for random number generation to ensure reproducibility.}
}
\value{
A ggplot object showing the prior distributions for the specified
component and type.
}
\description{
Plotting prior distributions as specified in the different model
components can help to identify accidental prior misspecification and to
understand the implications of specific prior choices.
}
\examples{
noise_model <- noise_dPCR()
plot_prior(noise_model, type = "partitions")
}
