% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{positive_partitions}
\alias{positive_partitions}
\title{Model the number of positive partitions in a dPCR assay}
\usage{
positive_partitions(
  measurements = NULL,
  id_col = "sample_id",
  replicate_col = NULL,
  positive_partitions_col = "positive_partitions",
  total_partitions_col = NULL,
  n_averaged = 1,
  n_averaged_col = NULL
)
}
\arguments{
\item{measurements}{A \code{data.frame} with each row representing one dPCR
measurement. Must have at least a column with sample ids and columns for
the numbers of total and positive partitions.}

\item{id_col}{Name of the column containing the id of the sample.}

\item{replicate_col}{Name of the column containing the replicate ID of each
measurement. This is used to identify multiple measurements (biological
replicates) of the same sample. Should be \code{NULL} if only one biological
replicate per sample was made.}

\item{positive_partitions_col}{Name of the column in the \code{measurements}
data.frame containing the number of positive partitions (e.g. positive
droplets for ddPCR) in the dPCR reaction of each measurement. If several
technical replicates are used, this should be the AVERAGE number of
positive partitions per replicate.}

\item{total_partitions_col}{Name of the column in the \code{measurements}
data.frame containing the number of total partitions (e.g. valid droplets
for ddPCR) in the dPCR reaction of each measurement. If several technical
replicates are used, this should be the AVERAGE number of valid partitions
per replicate. Note that this is really the number of \emph{valid} partitions,
not the number of positive partitions.}

\item{n_averaged}{The number of technical replicates (i.e. repeated PCR runs)
used for each biological replicate. The concentration provided in the
\code{measurements} \code{data.frame} is assumed to be the average or pooled estimate
from several technical replicates. Can be either a single number (it is
then assumed that the number of averaged replicates is the same for each
observation) or a vector (one value for each observation).}

\item{n_averaged_col}{Name of the column in the \code{measurements} data.frame
containing the number of technical replicates over which the measurements
have been averaged/pooled. This is an alternative to specifying
\code{n_averaged}.}
}
\description{
This option will fit a binomial regression model to the number
of positive partitions in a dPCR assay.
}
