% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCRfit.R
\name{predict.dPCRfit_result}
\alias{predict.dPCRfit_result}
\title{Predict method for dPCR Model Fits}
\usage{
\method{predict}{dPCRfit_result}(
  object,
  newdata,
  interval = c("none", "confidence", "prediction"),
  keep_data = FALSE
)
}
\arguments{
\item{object}{Object of class "dPCRfit_result"}

\item{newdata}{An optional data frame in which to look for covariates with
which to predict. If omitted, the fitted values are used.}

\item{interval}{Type of interval calculation. Can be abbreviated. Currently,
only "none" and "confidence" are supported. "None" will only give the mean
response. Confidence will give various summary statistics of the response.
Prediction intervals (i.e. including dPCR measurement noise) will be
added in the future.}

\item{keep_data}{Logical. If TRUE, the newdata is returned together with the
predictions.}
}
\value{
A data.table with the predicted values. If keep_data is TRUE, the
covariates are included in this data.table.
}
\description{
Predict values based on a fitted dPCR-specific regression model.
}
