% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{print.model_component}
\alias{print.model_component}
\title{Print method for model_component}
\usage{
\method{print}{model_component}(x, ...)
}
\arguments{
\item{x}{A model_component object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print method for model_component
}
\keyword{internal}
