% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{set_fit_opts}
\alias{set_fit_opts}
\title{Configure the model fitting}
\usage{
set_fit_opts(sampler = sampler_stan_mcmc(), model = model_stan_opts())
}
\arguments{
\item{sampler}{Which type of sampler should be used for model fitting?
Currently, only [sampler_stan_mcmc()] is supported.}

\item{model}{Details about the model file to be used, see
[model_stan_opts()].}
}
\value{
A `list` with the model fitting options.
}
\description{
Sets model fitting options in `dPCRfit`.
}
