% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_concentrations_dPCR}
\alias{sim_concentrations_dPCR}
\title{Simulate concentration measurements based on the dPCR partition count model}
\usage{
sim_concentrations_dPCR(
  n_samples = 1,
  lambda,
  m_partitions,
  conversion,
  prePCR_cv,
  n_replicates = 1,
  distribution = "gamma",
  get_partitions = FALSE
)
}
\arguments{
\item{n_samples}{How many measurements of the sample do you want to simulate?}

\item{lambda}{Expected concentration in the sample}

\item{m_partitions}{Number of partitions (vector of length n_replicates or
scalar)}

\item{conversion}{Conversion factor, i.e. volume v * scaling factor s}

\item{prePCR_cv}{Pre-PCR coefficient of variation}

\item{n_replicates}{Number of replicates}

\item{distribution}{Pre-PCR variation distribution, either "gamma" or "lnorm"}

\item{get_partitions}{If TRUE, return a data frame with concentration,
positive partitions and total partitions}
}
\value{
A vector of simulated measurements
}
\description{
Simulate concentration measurements based on the dPCR partition count model
}
