% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_partitions}
\alias{sim_partitions}
\title{Simulate number of valid partitions across dPCR runs}
\usage{
sim_partitions(
  n_samples,
  n_replicates,
  max_partitions,
  partition_loss_logit_mean,
  partition_loss_logit_sd,
  partition_loss_max
)
}
\arguments{
\item{n_samples}{Number of different samples}

\item{n_replicates}{Number of replicates per sample}

\item{max_partitions}{Maximum number of partitions}

\item{partition_loss_logit_mean}{Mean proportion of invalid partitions
(logit-level)}

\item{partition_loss_logit_sd}{Sd of proportion of invalid partitions
(logit-level)}

\item{partition_loss_max}{Upper bound for partition loss proportion}
}
\value{
Simulated partition numbers, structured as a list with one element
per sample, where each element is a vector giving the partition numbers
across replicates
}
\description{
Simulate number of valid partitions across dPCR runs
}
