% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCRfit.R
\name{summary.dPCRfit_result}
\alias{summary.dPCRfit_result}
\title{Provide a summary of a dPCR Model Fit}
\usage{
\method{summary}{dPCRfit_result}(object, ...)
}
\description{
Summarizes a dPCR Model Fit, including coefficient estimates and
model diagnostics.
}
