% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{concentration_measurements}
\alias{concentration_measurements}
\title{Model concentration measurements}
\usage{
concentration_measurements(
  measurements = NULL,
  id_col = "sample_id",
  concentration_col = "concentration",
  replicate_col = NULL,
  n_averaged = 1,
  n_averaged_col = NULL,
  total_partitions_col = NULL,
  distribution = "gamma"
)
}
\arguments{
\item{measurements}{A \code{data.frame} with each row representing one dPCR
measurement. Must have at least a column with sample ids and a column with
concentration measurements.}

\item{id_col}{Name of the column containing the id of the sample.}

\item{concentration_col}{Name of the column containing the measured
concentrations.}

\item{replicate_col}{Name of the column containing the replicate ID of each
measurement. This is used to identify multiple measurements made of a
sample from the same date. Should be \code{NULL} if only one measurement per
sample was made.}

\item{n_averaged}{The number of replicates over which the measurements have
been averaged. This is typically used as an alternative to providing
several replicates per sample (i.e. the concentration provided in the
\code{measurements} \code{data.frame} is the average of several replicates).
Can be either a single number (it is then assumed that the number of
averaged replicates is the same for each observation) or a vector (one
value for each observation).}

\item{n_averaged_col}{Name of the column in the \code{measurements} data.frame
containing the number of replicates over which the measurements have been
averaged. This is an alternative to specifying \code{n_averaged}.}

\item{total_partitions_col}{Name of the column in the \code{measurements}
data.frame containing the total number of partitions (e.g. droplets for
ddPCR) in the dPCR reaction of each measurement. Only applies to
concentration measurements obtain via dPCR. Can be used by the
\code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}} and \code{\link[=LOD_estimate_dPCR]{LOD_estimate_dPCR()}} modeling components. Note
that this is really the
\emph{total} number of partitions, not just the number of positive partitions.}

\item{distribution}{Parametric distribution for concentration measurements.
Currently supported are "gamma" (default and recommended), "log-normal",
"truncated normal", and "normal". The "truncated normal" and "normal"
options are not recommended for use in practice.}

\item{composite_window}{Over how many days has each measured sample been
collected? If 1 (default), samples represent single days. If larger than 1,
samples are assumed to be equivolumetric composites over several dates. In
this case, the supplied dates represent the last day included in each
sample.}
}
\description{
Model concentration measurements
}
