% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{noise_}
\alias{noise_}
\title{Model measurement noise (internal helper function)}
\usage{
noise_(
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  cv_type = "constant_cv",
  total_partitions_prior_mu = NULL,
  total_partitions_prior_sigma = NULL,
  total_partitions_observe = NULL,
  partition_variation_prior_mu = NULL,
  partition_variation_prior_sigma = NULL,
  volume_scaled_prior_mu = NULL,
  volume_scaled_prior_sigma = NULL,
  prePCR_noise_type = "log-normal",
  use_taylor_approx = TRUE,
  modeldata
)
}
\arguments{
\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements.}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{cv_type}{One out of "constant" (default), "constant_var", or "dPCR". If
"constant", the coefficient of variation is estimated as a constant/single
parameter for all observations. If "dPCR", the coefficient of variation is
modeled as a function of the expected concentration according to the
statistical properties of dPCR. In particular, this model predicts a higher
coefficient of variation at smaller concentrations, which often leads to a
better model fit. If "constant_var", not the coefficient of variation but
the variance of measurements is modeled as constant. This is usually a
misspecification and is only supported for comparison purposes.}

\item{total_partitions_prior_mu}{Prior (mean) on the total number of
partitions in the dPCR reaction.}

\item{total_partitions_prior_sigma}{Prior (standard deviation) on the total
number of partitions in the dPCR reaction. If this is set to zero, the
total number of partitions will be fixed to the prior mean and not
estimated.}

\item{total_partitions_observe}{If TRUE, the total number of partitions is
taken from the supplied measurements `data.frame`. This requires that the
argument `total_partitions_col` is specified in [concentrations_observe()].}

\item{partition_variation_prior_mu}{Prior (mean) on the coefficient of
variation of the total number of partitions in the dPCR reaction. Usually,
the maximum number of partitions possible for a given dPCR chip is not
reached, i.e. a certain number of partitions is lost. This loss varies
between PCR runs, and is modeled as log-normal distributed in EpiSewer.}

\item{partition_variation_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of the total number of partitions in the dPCR
reaction. If this is set to zero, the partition variation will be fixed to
the prior mean and not estimated.}

\item{volume_scaled_prior_mu}{Prior (mean) on the conversion factor
(partition volume multiplied with the scaling of concentration in the
assay) for the dPCR reaction. See details for further explanation.}

\item{volume_scaled_prior_sigma}{Prior (standard deviation) on the conversion
factor (partition volume multiplied with the scaling of concentration in
the assay) for the dPCR reaction. If this is set to zero, the conversion
factor will be fixed to the prior mean and not estimated.}

\item{prePCR_noise_type}{The parametric distribution to assume for noise
before the PCR assay. Currently supported are "log-normal" and "gamma". The
choice of the parametric distribution typically makes no relevant
difference for the noise model, but can make a relevant difference for the
LOD model if [LOD_estimate_dPCR()] is used.}

\item{use_taylor_approx}{If TRUE (default), a Taylor expansion approximation
is used to estimate the CV of measurements under pre-PCR noise. The
approximation is very accurate, unless concentrations are extremely high
(so high that the quality of the measurements from dPCR would anyway be
questionable).}
}
\description{
This helper function is called from specific noise modeling
  function. [noise_constant_cv()] is a constant coefficient of variation
  model, [noise_constant_var()] is a constant variance model, and
  [noise_dPCR()] is a noise model specialized for digital PCR (`cv_type =
  "dPCR"`).
}
\keyword{internal}
