% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dPCR.R
\name{nondetect_dPCR}
\alias{nondetect_dPCR}
\title{Model non-detects in digital PCR data}
\usage{
nondetect_dPCR(drop_prob = 1e-10)
}
\arguments{
\item{drop_prob}{Probability for non-detection below which likelihood
contributions of observed concentrations are dropped. This avoids numerical
issues of the non-detection model at high concentrations (very small
non-detection probabilities) that can otherwise affect sampling speed.
Since these likelihood contributions will be virtually zero for almost all
samples anyway, parameter estimates are practically not affected.}
}
\description{
Concentrations below a certain threshold may not be detectable
  and thus erroneously measured as 0. This option adjusts for a
  non-detects based on the statistical properties of digital PCR (dPCR)
  and includes zero measurements in the likelihood.

In effect, zero measurements provide a signal that the
  concentration in the respective sample was likely below the limit of
  detection, but we don't know what the exact concentration was.
}
\details{
Non-detects are modeled using a hurdle model. The model
  uses the number of partitions in the dPCR reaction and the conversion
  factor as defined and estimated by [noise_estimate_dPCR()]. It can
  therefore only be used together with `noise = noise_estimate_dPCR()` in
  [concentration_measurements()].
}
\concept{{LOD models}}
